refactor validation-rules.js 🥝 ..
